The issue is actually related to the fact that Tomcat is by default configured with maxSwallowSize=2 MB, so we need to only test with Tomcat.
